% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gen_interval.R
\name{extract_gen_interval}
\alias{extract_gen_interval}
\title{Extract generation interval information from text lines of Retriever output.}
\usage{
extract_gen_interval(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted generation interval information.
}
\description{
This function extracts generation interval information from text lines based on the specified language.
}
\examples{
lines <- c("GENERATION INTERVAL",
           " ", " "," "," ",
           "2000 5 6 5.5",
           "2001 5.5 6.5 6",
           "2002 6 7 6.5",
           " ", " "," "," ",
           "PEDIGREE COMPLETENESS")
extract_gen_interval(lines, "ENG")

}
