% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inbreeding.R
\name{extract_inbreeding}
\alias{extract_inbreeding}
\title{Extract inbreeding information from text lines of Retriever output.}
\usage{
extract_inbreeding(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted inbreeding information.
}
\description{
This function extracts inbreeding information from text lines based on the specified language.
}
\examples{
lines <- c("INBREEDING",
           " ", " "," "," "," ",
           "2000 0.10 0.05 0.06 0.03 0.02 0.01",
           "2001 0.12 0.06 0.07 0.04 0.03 0.02",
           "2002 0.15 0.08 0.09 0.05 0.04 0.03",
           " "," "," "," ")
extract_inbreeding(lines, "ENG")

}
