% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mean_age_parents.R
\name{extract_mean_age_parents}
\alias{extract_mean_age_parents}
\title{Extract mean age of parents information from text lines of Retriever output.}
\usage{
extract_mean_age_parents(lines, language, parent)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}

\item{parent}{A character string ("fathers" for sires or "mothers" for dams).}
}
\value{
A data frame containing the extracted mean age of parents information with columns:
\itemize{
\item Year of birth
\item Average age of parents (in years)
}
}
\description{
This function extracts the mean age of parents information from text lines based on the specified language.
}
\examples{
lines <- c(
  "AGE fathers",
  "    ",
  "________________________________________________________",
  " born                    age",
  " year         1       2       3       4       5       6       7       8       9      10",
  "________________________________________________________",
  "2000        10      12      8       10      15      7       10      8       10      10",
  "2001        9       11      10      12      14      9       10      8       9       8",
  "2002        11      10      9       13      12      8       11      7       10      9",
  "         ",
  "________________________________________________________"
)
extract_mean_age_parents(lines, "ENG", parent = "fathers")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
