% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parent_contribution.R
\name{extract_parent_contribution}
\alias{extract_parent_contribution}
\title{Extract parent contribution information from text lines of Retriever output.}
\usage{
extract_parent_contribution(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted parent contribution information.
}
\description{
This function extracts parent contribution information from text lines based on the specified language.
}
\examples{
lines <- c("Top sires and their contribution per year",
           " ", " ", " ", " ", " ",
           "2000 5 0.2 0.15 0.1 0.08 0.07 0.05 0.04 0.03 0.02 0.01",
           "2001 5 0.22 0.17 0.12 0.1 0.08 0.06 0.05 0.03 0.02 0.01",
           "2002 6 0.25 0.2 0.15 0.12 0.1 0.08 0.07 0.05 0.03 0.02",
           " ", " ", " ",
           "INBREEDING")
extract_parent_contribution(lines, "ENG")

}
