% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parent_littersize.R
\name{extract_parent_littersize}
\alias{extract_parent_littersize}
\title{Extract parent and litter size information from text lines of Retriever output.}
\usage{
extract_parent_littersize(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted parent and litter size information.
}
\description{
This function extracts parent and litter size information from text lines based on the specified language.
}
\examples{
lines <- c(
  "Number of parents and litter size per year",
  "________________________________________________________",
  "Y of   number   calves/   number   litter   calves/ ",
  "birth   litter   litter   fathers  /father   father ",
  "________________________________________________________",
  "2000       100      2.0       50      2.0      1.0",
  "2001       110      2.5       55      2.2      1.1",
  "2002       120      3.0       60      2.5      1.2",
  "________________________________________________________"
)

extract_parent_littersize(lines, "ENG")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
