% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pedigree_completeness.R
\name{extract_pedigree_completeness}
\alias{extract_pedigree_completeness}
\title{Extract pedigree completeness information from text lines of Retriever output.}
\usage{
extract_pedigree_completeness(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted pedigree completeness information.
}
\description{
This function extracts pedigree completeness information from text lines based on the specified language.
}
\examples{
lines <- c("PEDIGREE COMPLETENESS",
           " ", " ", " ", " ", " ",
           "2000 2.5 80 65 50 40 30",
           "2001 2.7 85 70 55 45 35",
           "2002 3.0 90 75 60 50 40",
           " ", " ",
           "Top sires and their contribution per year")
extract_pedigree_completeness(lines, "ENG")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
