% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pop_size.R
\name{extract_pop_size}
\alias{extract_pop_size}
\title{Extract population size information from text lines of Retriever output.}
\usage{
extract_pop_size(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted population size information.
}
\description{
This function extracts population size information from text lines based on the specified language.
}
\examples{
lines <- c("POPULATIONSIZE: number of calves born per year",
           " ", " "," "," ",
           "2000 10 20 15 5 30 50",
           "2001 15 25 20 10 35 60",
           "2002 20 30 25 15 40 70",
            " "," "," ",
           "total number of animals in pedigree")
extract_pop_size(lines, "ENG")

}
