% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_section.R
\name{extract_section}
\alias{extract_section}
\title{Extract a section from a character vector}
\usage{
extract_section(
  lines,
  start_with,
  ends_with,
  force_first_grep_start = F,
  force_first_grep_end = F,
  skip_initial_n_lines,
  skip_last_n_lines,
  column_names = NA,
  fill_colnames = NA,
  skipcols = NA,
  exclude_between_char = NA
)
}
\arguments{
\item{lines}{A character vector containing the lines of text to use as input.}

\item{start_with}{A character string indicating the phrase marking the start of the section.}

\item{ends_with}{A character string indicating the phrase marking the end of the section.}

\item{force_first_grep_start}{Logical (default = FALSE). If TRUE, only the first match of \code{start_with} will be considered.}

\item{force_first_grep_end}{Logical (default = FALSE). If TRUE, only the first match of \code{ends_with} after the start match will be considered.}

\item{skip_initial_n_lines}{Numeric. The number of lines to skip after the start match.}

\item{skip_last_n_lines}{Numeric. The number of lines to skip before the end match.}

\item{column_names}{Optional. A character vector specifying column names for the resulting data frame.}

\item{fill_colnames}{Optional. A character string specifying how to fill column names in specific situations (e.g., "parents_age").}

\item{skipcols}{Optional. A numeric vector specifying column indices to skip.}

\item{exclude_between_char}{Optional. A character string indicating characters between which content should be excluded.}
}
\value{
A data frame representing the extracted section from the input.
}
\description{
This function extracts a section from a character vector based on specified start and end phrases, skipping initial and final lines, and optionally excluding content between specified characters.
}
\examples{
lines <- c("GENERATION INTERVAL  ",
           " some text to skip", "some text to skip ",
           "",  "", "", # empty lines are skipped automatically
           "2000 5 6 5.5",
           "2001 5.5 6.5 6",
           "2002 6 7 6.5",
           "",  "", "", # empty lines are skipped automatically
           "some text to skip ", "some text to skip ",
           "PEDIGREE COMPLETENESS    ")
lines
extract_section(lines, start_with = "GENERATION INTERVAL",
           ends_with = "PEDIGREE COMPLETENESS", skip_initial_n_lines = 3,
           skip_last_n_lines = 3)

}
