% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crossref_table.R
\name{get_crossref_table}
\alias{get_crossref_table}
\title{Get Cross-Reference Table}
\usage{
get_crossref_table()
}
\value{
A data frame with three columns: "lang" (language), "section" (section name), and "keyword" (corresponding keyword).
}
\description{
This function builds and returns a cross-reference table mapping section names and language to keywords
}
\examples{
crossref_table <- get_crossref_table()

}
