% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_linepoint_plot.R
\name{make_linepoint_plot}
\alias{make_linepoint_plot}
\title{Make a line plot with points with year of birth on the x-axis}
\usage{
make_linepoint_plot(data)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. First column must be named 'Year'.}
}
\value{
A ggplot object representing the line plot with points.
}
\description{
This function generates a line plot with line-points for each column in the input data and year of birth in the x-axis
}
\examples{
data <- data.frame(Year = c(2000, 2001, 2002),
                   Group1 = c(10, 20, 15),
                   Group2 = c(5, 10, 8))
make_linepoint_plot(data)

}
