% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stacked_barplot.R
\name{make_stacked_barplot}
\alias{make_stacked_barplot}
\title{Create a stacked bar plot}
\usage{
make_stacked_barplot(data, levels_order)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. The first column is used as x-axis.}

\item{levels_order}{A vector specifying the order of levels for stacking.}
}
\value{
A ggplot object representing the stacked bar plot.
}
\description{
This function generates a stacked bar plot based on the provided data and the specified order of levels.
}
\examples{
data <- data.frame(Year = c(2000, 2001, 2002),
                   Group1 = c(10, 20, 15),
                   Group2 = c(5, 10, 8))
levels_order <- c("Group1", "Group2")
make_stacked_barplot(data, levels_order)

}
