% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stacked_lineplot.R
\name{make_stacked_lineplot}
\alias{make_stacked_lineplot}
\title{Create a stacked area plot}
\usage{
make_stacked_lineplot(data)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. First column must be named 'Year'.}
}
\value{
A ggplot object representing the stacked area plot.
}
\description{
This function generates a stacked area plot based on the provided data.
}
\examples{
data <- data.frame(Year = c(2000, 2001, 2002),
                   Group1 = c(10, 20, 15),
                   Group2 = c(5, 10, 8))
make_stacked_lineplot(data)

}
