% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_retriever.R
\name{process_retriever}
\alias{process_retriever}
\title{Process Retriever output log file}
\usage{
process_retriever(
  file_name,
  language,
  verbose = F,
  xinterval = NULL,
  delta_intervals = NULL
)
}
\arguments{
\item{file_name}{A character string with the full path of the Retriever log file.}

\item{language}{A character string specifying the language of the log file ("DUT" for Dutch or "ENG" for English).}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages. Default is \code{FALSE}.}

\item{xinterval}{Numeric. A numeric vector of length 2 specifying the x-axis limits for plots. It can be used to focus on specific years intervals. Default is \code{NULL}, i.e., plots show all years.}

\item{delta_intervals}{Numeric. A numeric vector specifying the intervals for which to compute the annual and generational deltas for inbreeding and kinship. Default is \code{NULL}.}
}
\value{
A list containing processed data (as data frames) and plots for population size, inbreeding, generation intervals, mean age of parents, pedigree completeness, litter size, and top sires' contribution.
}
\description{
This function processes a Retriever output log file, extracting various metrics and generating plots based on the content.
}
\examples{
# Process a Retriever log file in English
# results <- process_retriever("path/to/retriever.log", "ENG")

# Access processed data and plots
# str(results, max.level = 1, give.attr = F) # look into the extracted data frames and plots
# results$pop_size # Population size data
# results$pop_size_plot # Population size plot

}
