% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_extract.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely Execute an Extraction Function with Error Logging}
\usage{
safe_extract(function_to_exec, ..., return_error = TRUE)
}
\arguments{
\item{function_to_exec}{A function to be executed within a \code{tryCatch} block.}

\item{...}{Additional arguments passed to \code{function_to_exec}.}

\item{return_error}{Logical; if \code{TRUE}, errors and warnings are logged in \code{error_log}.}
}
\value{
The result of \code{function_to_exec(...)} if successful, otherwise \code{NULL} if error is detected.
If a warning or message is detected, the results of \code{function_to_exec()} are still returned.
}
\description{
This function attempts to execute a given function and captures any errors
or warnings that occur during execution. Errors and warnings are logged in an \code{error_log}
variable located in the parent environment.
}
\examples{
error_log <- list()  # Initialize an error log
safe_extract(log, -1)  # Example usage with a function that may generate warnings/errors

}
