% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_labrador.R
\name{theme_labrador}
\alias{theme_labrador}
\title{Custom theme for Labrador package}
\usage{
theme_labrador()
}
\value{
A ggplot2 theme object.
}
\description{
This function defines a custom theme for use in plots within the Labrador package.
}
