% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inbreeding.R
\name{extract_inbreeding}
\alias{extract_inbreeding}
\title{Extract inbreeding information from text lines of Retriever output.}
\usage{
extract_inbreeding(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text. Supported values are:
\itemize{
\item \code{"DUT"} for Dutch
\item \code{"ENG"} for English
}}
}
\value{
A data frame containing extracted inbreeding and kinship information with the following columns:
\itemize{
\item \code{Year} -  The year of birth of animals.
\item \code{F_all_animals} - The average inbreeding coefficient for all individuals born in a year.
\item \code{f_inc.self} - The average kinship coefficient (including selfing) for all individuals born in a year.
\item \code{f_exc.self} - The average kinship coefficient (excluding selfing) for all individuals born in a year.
\item \code{f_parents} - The average kinship coefficient (excluding selfing) for animals born in a year, which later become a parent.
\item \code{f_sires} - The average kinship coefficient (excluding selfing) for animals born in a year, which later become a sire.
\item \code{f_dams} - The average kinship coefficient (excluding selfing) for animals born in a year, which later become a dam.
}
}
\description{
This function extracts inbreeding information from text lines based on the specified language.
}
\examples{
# Example input lines (simulating a Retriever log file with inbreeding data)
lines <- c(
  "INBREEDING",
  "________________________________________________________",
  "  Yr.of     F       f       f       f       f     f    ",
  "  birth    all     incl.  excl. parents fathers mothers",
  "    -    animals   self   self        ",
  "________________________________________________________",
  "2000    0.10    0.05    0.06    0.03    0.02    0.01",
  "2001    0.12    0.06    0.07    0.04    0.03    0.02",
  "2002    0.15    0.08    0.09    0.05    0.04    0.03",
  "________________________________________________________"
)
extract_inbreeding(lines, "ENG")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
