% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mean_age_parents.R
\name{extract_mean_age_parents}
\alias{extract_mean_age_parents}
\title{Extract mean age of parents information from text lines of Retriever output.}
\usage{
extract_mean_age_parents(lines, language, parent)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}

\item{parent}{A character string ("fathers" for sires or "mothers" for dams).}
}
\value{
A data frame representing the extracted mean age of parents information.
}
\description{
This function extracts the mean age of parents information from text lines based on the specified language.
}
\examples{
lines <- c("AGE fathers",
           " ", " "," "," ",
           "2000 30 25",
           "2001 32 26",
           "2002 35 28",
           " ", " "," "," ",
           "AGE mothers")
extract_mean_age_parents(lines, "ENG", parent = "fathers")

}
