% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parent_contribution.R
\name{extract_parent_contribution}
\alias{extract_parent_contribution}
\title{Extract parent contribution information from text lines of Retriever output.}
\usage{
extract_parent_contribution(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted parent contribution information.
}
\description{
This function extracts parent contribution information from text lines based on the specified language.
}
\examples{
lines <- c(
  "Top sires and their contribution per year",
  paste0("________________________________________________________________________________",
         "_________________________________"),
  paste0("|                  the Topsire             |                   contribution of top 10",
         " sires (\%)         |"),
  paste0("  year #Sires   | name                    id               |    1     2     3     4",
         "     5     6     7     8     9    10 |"),
  paste0("________________________________________________________________________________",
         "_________________________________"),
  paste0("2000       5    | TopSireA                12345            |  20    15    12    10",
         "     9     8     7     6     7     6 |"),
  paste0("2001       5    | TopSireB                23456            |  22    17    14    10",
         "     9     8     7     6     4     3 |"),
  paste0("2002       6    | TopSireC                34567            |  25    20    15    12",
         "    10     8     5     3     2     2 |"),
  "________________________________________________________"
)
extract_parent_contribution(lines, "ENG")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
