% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pedigree_completeness.R
\name{extract_pedigree_completeness}
\alias{extract_pedigree_completeness}
\title{Extract pedigree completeness information from text lines of Retriever output.}
\usage{
extract_pedigree_completeness(lines, language)
}
\arguments{
\item{lines}{A character vector containing the lines of text from Retriever.}

\item{language}{A character string specifying the language of the text ("DUT" for Dutch or "ENG" for English).}
}
\value{
A data frame representing the extracted pedigree completeness information.
}
\description{
This function extracts pedigree completeness information from text lines based on the specified language.
}
\examples{
lines <- c(
"PEDIGREE COMPLETENESS",
"________________________________________________________________________",
"Y of   average      | percentage animals with # of generations ",
"birth  generation   | in pedigree completely known",
"equivalent   |    0       1       2       3       4    5 or more  ",
"________________________________________________________________________",
"2000       2.5    20.0    15.0    25.0    20.0    10.0    10.0",
"2001       2.7    18.0    17.0    24.0    21.0    10.0    10.0",
"2002       3.0    15.0    20.0    22.0    23.0    10.0    10.0",
"________________________________________________________________________"
)

extract_pedigree_completeness(lines, "ENG")

}
\seealso{
\code{\link{extract_text}}, \code{\link{extract_section}}
}
