% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_section.R
\name{extract_section}
\alias{extract_section}
\title{Extract and Process a Section of Data}
\usage{
extract_section(
  lines,
  exclude_between_char = NA,
  trim_asterisks = F,
  skipcols = NA,
  column_names = NA,
  fill_colnames = NA,
  fixed_col_width = NA
)
}
\arguments{
\item{lines}{A character vector containing lines of text to be processed.}

\item{exclude_between_char}{A character indicating a delimiter (e.g., \code{"_"}) for removing text enclosed between consecutive occurrences of the character.}

\item{trim_asterisks}{Logical (default \code{FALSE}). If \code{TRUE}, removes all asterisks (\code{*}) from the text.}

\item{skipcols}{A numeric vector specifying column indices to be removed from the extracted data.}

\item{column_names}{A character vector of column names. If provided, it must match the number of extracted columns.}

\item{fill_colnames}{A special keyword (e.g., \code{"parents_age"}) to automatically generate column names.}

\item{fixed_col_width}{A numeric vector specifying the fixed column widths for parsing the data. If \code{NA}, whitespace-based splitting is used instead.}
}
\value{
A data frame with processed numeric values and appropriately assigned column names.
}
\description{
This function extracts tabular data from a list of lines, processes column names,
removes unwanted content, and converts extracted data into a numeric data frame.
}
\details{
\itemize{
\item The function automatically splits lines into columns based on whitespace.
\item All extracted values are converted to numeric.
\item If \code{exclude_between_char} is provided, content between matching symbols is removed.
\item If \code{trim_asterisks = TRUE}, all \code{*} symbols are stripped.
\item If \code{fill_colnames = "parents_age"}, the function auto-generates column names with \code{Year} followed by \code{Age_1}, \code{Age_2}, etc.
\item If \code{column_names} is provided, its length must match the number of extracted columns, otherwise an error is thrown.
\item If \code{fixed_col_width} do not count leading spaces for first col which hare removed.
}
}
\examples{
lines <- c("2020 25 30 35", "2021 26 31 36")
extract_section(lines, column_names = c("Year", "Age_A", "Age_B", "Age_C"))

}
